libs/platform_shared/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "platform-shared"
version = "0.1.0"
description = "Shared config and utilities for producer/worker services"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "sqlalchemy>=2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["platform_shared*"]
